Adding new lang plugins, making overwrite optional
Decided that I don't always want to make `init_vim` overwrite the
prebuilt vimrc and making it optional. Also added support for more
languages that I do use in development.

+++ init_vim.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-echo "Overriding VIMRC with the local version"
-
-cp .vimrc ~/
+if [[ ! -f ~/.vimrc ]] || [[ $1 == '-f' ]]; then
+    echo "Overriding VIMRC with the local version"
+    cp .vimrc ~/
+fi
 
 ./install_core_plugins.sh
 ./install_lang_plugins.sh
